fix: Symlinks to non-existent directories, symlinks re-creating

+++ aws_lambda_builders/utils.py
@@ -198,6 +198,9 @@ def create_symlink_or_copy(source: str, destination: str) -> None:
     """Tries to create symlink, if it fails it will copy source into destination"""
     LOG.debug("Creating symlink; source: %s, destination: %s", source, destination)
     try:
+        if Path(destination).exists() and Path(destination).is_symlink():
+            # The symlink is already in place, don't try re-creating it
+            return
         os.symlink(Path(source).absolute(), Path(destination).absolute())
     except OSError as ex:
         LOG.warning(
